requirements.txt
pyobjc_core==9.0
